Moving pages to routes folder
